fix convention and rename type action, reducer
