+++ src/app/pages/Cart/partials/SectionCart.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useSelector } from 'react-redux';
 import CartList from './CartList';
 import TotalList from './TotalList';
-import { ICart } from './../../../shared/interfaces/cart';
+import { ICart } from '../../../shared/interfaces/cart';
 import { formatFixed } from '../../../shared/common/common';
 
 const SectionCart = () => {
