+++ src/app/redux/appReducer.ts
@@ -1,6 +1,6 @@
 import React from 'react';
 import { combineReducers } from 'redux';
-import cartReducer from './../pages/Cart/cart.reducer';
+import cartReducer from '../pages/cart/cart.reducer';
 
 const rootReducer = combineReducers({
   cartReducer: cartReducer
